Move more utils from snap controller tests
