+++ packages/controllers/src/test-utils/index.ts
@@ -1,1 +1,2 @@
+export * from './controller';
 export * from './execution-environment';
